requirements.txt
# Core library for document parsing
docling

# Library for PDF metadata extraction
PyMuPDF

# Libraries for the RAG system
langchain
langchain-community
langchain-groq
langchain-huggingface  # Nova biblioteca
faiss-cpu
python-dotenv